Move Twist and Battery stuff into a Model-View system
The basic idea is separate the data handling and the display/user
handling parts of the code. The data and ROS handling goes into Models
that encapsulate the data reading and writing. These can then be bound
to views that display the data and handle user interaction.

This effort started in the avalon_web repository, but we are moving it
here now. These are the relevant commit messages from that effort:

    Convert the driving inputs to Views that bind to Twist

    This converts all of the Twist handling stuff into a more modular
    MVC structure.

    The use of arrow function the "() => { }" syntax is because javascript
    is stupid about "this" and the arrow syntax fixes this. In an arrow
    function "this" is bound to the surrounding scope of the function, not
    the function's caller. This allows us to make these functions callbacks
    without having to do weird binding things.

    Apply MVC to the battery stuff.

    Shows application of the the MVC model for the data coming in from
    ROS and going to the UI.